+++ app/system/modules/captcha/webpack.config.js
@@ -1,0 +1,13 @@
+module.exports = [
+
+    {
+        entry: {
+            "captcha-interceptor": "./app/interceptor.js"
+        },
+        output: {
+            filename: "./app/bundle/[name].js",
+            library: "Captcha"
+        }
+    }
+
+];